src/backend/requirements.txt
bcrypt==4.0.1
cryptography==39.0.2
email-validator>=1.1.1,<2.0.0
fastapi==0.88.0
httpx==0.23.3
ipython==8.11.0
itsdangerous==2.1.2
Jinja2==3.1.2
passlib==1.7.4
pymongo==4.3.3
python-dateutil==2.8.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
requests==2.28.2
starlette==0.22.0
uvicorn==0.20.0